pyproject.toml

[tool.black]
line-length = 120

[tool.mypy]
files = "src/gradualelixir/"
disallow_any_unimported = true
ignore_missing_imports = true

[tool.isort]
known_third_party = "gradualelixir"
profile="black"

[tool.poetry]
name = "gradualelixir"
version = "0.1.0"
description = "A gradual type checker with runtime checks for a mini elixir language"
authors = ["damif94 <damianferencz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pygments-style-github = "^0.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
mypy = "^0.941"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
Pygments = "^2.11.2"

[tool.poetry.scripts]
gradualelixir = "src.gradualelixir.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
